exercise 3 + exercise 4 challenge loop and flow control

+++ src/main/java/session_15_oop_recap/challenge/loops_and_flow_control/Exercise3.java
@@ -15,7 +15,7 @@ public class Exercise3 {
         if (number < 0) {
             System.out.println("Factorial is not defined for negative numbers.");
         } else {
-            int factorial = 1;
+            long factorial = 1;
             for (int i = 1; i <= number; i++) {
                 factorial *= i;
             }
